rename to character entity
