+++ src/app/api/characters/[id]/route.ts
@@ -1,4 +1,3 @@
-import { Character } from '@/models/character'
 import {
   getCharacterById,
   updateCharacterById,
